requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
email-validator==2.2.0
geopy==2.4.1
itsdangerous==2.2.0
Werkzeug==3.0.3
